tweak

+++ Frank.Apps.StarMap/Services/Physics.cs
@@ -5,7 +5,7 @@ namespace Frank.Apps.StarMap.Services;
 
 public class Physics
 {
-    public Vector2 Trajectory(Vector2 origin, int instant, float launchAngle, float initialVelocity)
+    public Vector2 Trajectory(Vector2 origin, float instant, float launchAngle, float initialVelocity)
     {
 
         Vector2 vector2 = new Vector2(0, 0);
@@ -14,11 +14,11 @@ public class Physics
         vector2.Y = CalculateVerticalVelocity(instant, initialVelocity, launchAngle);
         return vector2;
     }
-    public static float CalculateHorizontalVelocity(int instant, float initialVelocity, float launchAngle)
+    public static float CalculateHorizontalVelocity(float instant, float initialVelocity, float launchAngle)
     {
         return initialVelocity * MathF.Cos(launchAngle) * instant;
     }
-    public static float CalculateVerticalVelocity(int instant, float initialVelocity, float launchAngle)
+    public static float CalculateVerticalVelocity(float instant, float initialVelocity, float launchAngle)
     {
         return initialVelocity * MathF.Sin(launchAngle) * instant - 0.5F * 9.81F * instant * instant;
     }
